+++ src/main/java/mr/springbootapi/exception/GenericException.java
@@ -1,0 +1,7 @@
+package mr.springbootapi.exception;
+
+public class GenericException extends RuntimeException{
+    public GenericException(String message) {
+        super(message);
+    }
+}
