+++ src/main/java/mr/springbootapi/exception/InvalidPriceException.java
@@ -1,0 +1,7 @@
+package mr.springbootapi.exception;
+
+public class InvalidPriceException extends RuntimeException{
+    public InvalidPriceException(String message) {
+        super(message);
+    }
+}
