+++ src/main/java/mr/springbootapi/exception/ItemNotFoundException.java
@@ -1,0 +1,7 @@
+package mr.springbootapi.exception;
+
+public class ItemNotFoundException extends RuntimeException{
+    public ItemNotFoundException(String message) {
+        super(message);
+    }
+}
